pyproject.toml
[tool.poetry]
name = "nuke_from_orbit"
version = "0.1.0"
description = "A distributed load testing tool for running api and browser based tests on Looker"
authors = ["Colin Pistell <colinpistell@google.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
selenium = "*"
realbrowserlocusts = "*"
locustio = "==0.14.6"
looker-sdk = "*"
jinja2 = "*"
backoff-utils = "*"
kubernetes = "*"
google-cloud-container = "^2.2.0"
google-cloud-build = "^3.0.0"
google-cloud-storage = "^1.33.0"
google-api-python-client = "^1.12.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "*"
flake8 = "*"
pytest-mock = "^3.3.1"

[tool.poetry.scripts]
nuke = "nuke_from_orbit.cli:nuke"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
